Move NameValueCollection tests to optional converters
